add twoStep in confirm

+++ src/main/java/com/ndb/auction/resolver/UserResolver.java
@@ -96,10 +96,19 @@ public class UserResolver extends BaseResolver implements GraphQLQueryResolver, 
     }
 
     @PreAuthorize("isAuthenticated()")
-    public String deleteAccount() {
-        UserDetailsImpl userDetails = (UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        String id = userDetails.getId();
-        return userService.deleteUser(id);
+    public String deleteAccount() {        
+        return "To delete your account, please withdraw all your assets from NDB Wallet. Please note deleting process is irreversible.";
+    }
+
+    @PreAuthorize("isAuthenticated()")
+    public String confirmDeleteAccount(String text) {
+        if (text.equals("delete")) {
+            UserDetailsImpl userDetails = (UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            String id = userDetails.getId();
+            return userService.deleteUser(id);
+        } else {
+            return "failed";
+        }
     }
 
 //    @PreAuthorize("isAuthenticated")
